+++ multiples/tests/multiples_tests.c
@@ -12,12 +12,12 @@ int cmp_uint_fast64_t(const void *elem1, const void *elem2)
     return (f > s) - (f < s);
 }
 
-void factors_test_case(int n, int *expected_factors, size_t expected_factor_count)
+void factors_test_case(int n, unsigned int *expected_factors, size_t expected_factor_count)
 {
     uint_fast64_t factors_buffer[expected_factor_count];
     int factor_count;
     factors(n, factors_buffer, expected_factor_count, &factor_count);
-    assert(factor_count == expected_factor_count);
+    assert(factor_count == (int)expected_factor_count);
     qsort(factors_buffer, factor_count, sizeof(factors_buffer[0]), cmp_uint_fast64_t);
     for (int i = 0; i < factor_count; i++)
     {
@@ -28,16 +28,16 @@ void factors_test_case(int n, int *expected_factors, size_t expected_factor_coun
 void test_factors()
 {
 
-    factors_test_case(1, (int[]){1}, 1);
-    factors_test_case(2, (int[]){1, 2}, 2);
-    factors_test_case(3, (int[]){1, 3}, 2);
-    factors_test_case(4, (int[]){1, 2, 4}, 3);
-    factors_test_case(5, (int[]){1, 5}, 2);
-    factors_test_case(6, (int[]){1, 2, 3, 6}, 4);
-    factors_test_case(7, (int[]){1, 7}, 2);
-    factors_test_case(8, (int[]){1, 2, 4, 8}, 4);
-    factors_test_case(9, (int[]){1, 3, 9}, 3);
-    factors_test_case(10, (int[]){1, 2, 5, 10}, 4);
+    factors_test_case(1, (unsigned int[]){1}, 1);
+    factors_test_case(2, (unsigned int[]){1, 2}, 2);
+    factors_test_case(3, (unsigned int[]){1, 3}, 2);
+    factors_test_case(4, (unsigned int[]){1, 2, 4}, 3);
+    factors_test_case(5, (unsigned int[]){1, 5}, 2);
+    factors_test_case(6, (unsigned int[]){1, 2, 3, 6}, 4);
+    factors_test_case(7, (unsigned int[]){1, 7}, 2);
+    factors_test_case(8, (unsigned int[]){1, 2, 4, 8}, 4);
+    factors_test_case(9, (unsigned int[]){1, 3, 9}, 3);
+    factors_test_case(10, (unsigned int[]){1, 2, 5, 10}, 4);
     uint_fast64_t factors_buffer[1];
     int factor_count;
     factors(2, factors_buffer, 1, &factor_count);
