+++ primes/sieve_of_eratosthenes.c
@@ -3,7 +3,7 @@
 
 void sieve_of_eratosthenes(bool *composites_buffer, size_t composites_buffer_size)
 {
-    for (int i = 0; i < composites_buffer_size; i++)
+    for (int i = 0; (unsigned int)i < composites_buffer_size; i++)
     {
         composites_buffer[i] = false;
     }
