+++ solvers/s0007.c
@@ -38,5 +38,5 @@ bool solve0007(char *buffer, size_t buffer_size)
         }
     }
     int ret = snprintf(buffer, buffer_size, "%d", result);
-    return ret >= 0 && ret < buffer_size;
+    return ret >= 0 && (unsigned int)ret < buffer_size;
 }
